feat: Introduce incident decision handling and gRPC support for Tier-2
- Added IncidentDecisionRequest and IncidentDecisionResponse structs for handling incident decisions.
- Implemented incident_decision_handler to process incoming requests and resolve incident statuses.
- Updated the API routes to include /incident/decision for the new functionality.
- Refactored Tier-2 forwarding to use gRPC instead of HTTP, including necessary changes in the forwarder and related modules.
- Introduced new protobuf definitions for Tier-2 signals and requests.
- Enhanced signal processing logic to accommodate new decision-making criteria based on severity and confidence.
- Improved code readability and consistency by applying various refactorings and optimizations across multiple modules.

+++ via-core/crates/via-bench/examples/benchmark.rs
@@ -74,7 +74,7 @@ fn main() {
         }
 
         // --- Tick Simulation ---
-        let batch = sim.tick(tick_ms as u64 * 1_000_000); // ns
+        let batch = sim.tick(tick_ms * 1_000_000); // ns
 
         // "Ingest" logs
         let mut batch_latency_sum = 0.0;
@@ -97,7 +97,7 @@ fn main() {
 
                 // Process Event
                 let result = profile.process(
-                    t as u64 * tick_ms, // timestamp
+                    t * tick_ms, // timestamp
                     &record.traceId,    // unique_id
                     latency,            // value
                 );
